cleaned up process manager and documented
