+++ src/services/process/processManagerRequest.ts
@@ -1,0 +1,7 @@
+export interface ProcessManagerRequest {
+	key: string;
+	command: [];
+	stdOut: any;
+	stdIn: any;
+	stdErr: any;
+}
